requirements.txt
## Runtime
httpx>=0.27.0
mcp>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
sse-starlette>=1.6.0
boto3>=1.34.0
strands-agents>=0.1.0

## OpenTelemetry — integrates with AgentCore Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
